+++ src/extension.ts
@@ -23,6 +23,7 @@ import { TemplateOnTypingFormatter } from './providers/onTypeFormattingEditProvi
 import { TemplateCodeActionProvider } from './providers/codeActionProvider';
 import { TemplateFoldingRangeProvider } from './providers/foldingRangeProvider';
 import { TemplateDocumentSemanticTokensProvider, tokenTypesLegend } from './providers/documentSemanticTokensProvider';
+import { TemplateCallHierarchyProvider } from './providers/callHierarchyProvider';
 import { LanguageData } from './language/static/languageData';
 import { Quests } from './language/quests';
 import { QuestLinter } from './diagnostics/questLinter';
@@ -65,6 +66,10 @@ export async function activate(context: ExtensionContext) {
 
         if (getOptions()['codeLens']['enabled']) {
             context.subscriptions.push(vscode.languages.registerCodeLensProvider(TEMPLATE_MODE, new TemplateCodeLensProvider(data.modules, quests)));
+        }
+
+        if (getOptions()['callHierarchy']['enabled']) {
+            context.subscriptions.push(vscode.languages.registerCallHierarchyProvider(TEMPLATE_MODE, new TemplateCallHierarchyProvider(quests)));
         }
 
         if (getOptions()['diagnostics']['enabled']) {